requirements.txt
tensorflow-cpu
tflearn
nltk
underthesea
numpy
chardet
python-dotenv
flask
flask_cors